requirements.txt
torch==1.8.0
clip
opencv-python
numpy
seaborn
pillow